+++ backends/carp_webservices/test/carp_protocol_service_test.dart
@@ -14,8 +14,8 @@ void main() {
   CarpApp app;
   CarpUser user;
   String ownerId, name;
+  StudyProtocol protocol, custom;
 
-  /// Setup CARP and authenticate.
   /// Runs once before all tests.
   setUpAll(() async {
     Settings().debugLevel = DebugLevel.DEBUG;
@@ -35,9 +35,32 @@ void main() {
       password: password,
     );
 
+    CANSProtocolService().configureFrom(CarpService());
     ownerId = CarpService().currentUser.accountId;
     name = 'test_protocol';
-    CANSProtocolService().configureFrom(CarpService());
+
+    // a very simple protocol
+    protocol = StudyProtocol(
+        ownerId: ownerId,
+        name: name,
+        description: 'Generated from carp_webservices unit test.')
+      ..addMasterDevice(Smartphone(roleName: 'smartphone'));
+
+    // a custom protocol
+    var customDevice = CustomProtocolDevice(roleName: 'Custom device');
+    custom = StudyProtocol(
+        ownerId: ownerId,
+        name: 'custom_test_protocol',
+        description:
+            'Custom protocol generated from carp_webservices unit test.');
+    custom.addMasterDevice(customDevice);
+    custom.addTriggeredTask(
+        ElapsedTimeTrigger(
+            sourceDeviceRoleName: customDevice.roleName,
+            elapsedTime: Duration(seconds: 0)),
+        CustomProtocolTask(
+            name: 'Custom device task', studyProtocol: '{"version":"1.0"}'),
+        customDevice);
   });
 
   /// Runs once after all tests.
@@ -55,13 +78,8 @@ void main() {
     test(
       '- add',
       () async {
-        StudyProtocol protocol = StudyProtocol(
-            ownerId: ownerId,
-            name: '$name-1.1',
-            description: 'Generated from carp_webservices unit test.')
-          ..addMasterDevice(Smartphone(roleName: 'smartphone'));
-
-        await CANSProtocolService().add(protocol, '1.1');
+        print(toJsonString(protocol));
+        await CANSProtocolService().add(protocol);
       },
     );
 
@@ -76,37 +94,14 @@ void main() {
         //   '{"version":1}',
         // );
 
-        var customDevice = CustomProtocolDevice(roleName: 'Custom device');
-
-        StudyProtocol protocol = StudyProtocol(
-            ownerId: ownerId,
-            name: '$name',
-            description: 'Generated from carp_webservices unit test.');
-        protocol.addMasterDevice(customDevice);
-        protocol.addTriggeredTask(
-            ElapsedTimeTrigger(
-                sourceDeviceRoleName: customDevice.roleName,
-                elapsedTime: Duration(seconds: 0)),
-            CustomProtocolTask(
-                name: 'Custom device task', studyProtocol: '{"version":"1.0"}'),
-            customDevice);
-
-        await CANSProtocolService().add(protocol, '1.3');
+        await CANSProtocolService().add(custom, '1.3');
       },
     );
 
     test(
       '- addVersion',
       () async {
-        StudyProtocol protocol =
-            await CANSProtocolService().createCustomProtocol(
-          ownerId,
-          name,
-          'Made from Dart unit test.',
-          '{"version":2}',
-        );
-
-        await CANSProtocolService().addVersion(protocol);
+        await CANSProtocolService().addVersion(custom);
       },
     );