+++ carp_mobile_sensing/lib/domain/study_protocol.dart
@@ -35,11 +35,7 @@ class CAMSStudyProtocol extends StudyProtocol {
   StudyProtocolDescription protocolDescription;
 
   /// The owner of this study.
-  ProtocolOwner owner;
-
-  /// The unique id of the owner.
-  @override
-  String get ownerId => (owner != null) ? owner.id : super.ownerId;
+  StudyProtocolReponsible responsible;
 
   /// The [masterDevice] which is responsible for aggregating and synchronizing
   /// incoming data. Typically this phone.
@@ -47,12 +43,14 @@ class CAMSStudyProtocol extends StudyProtocol {
 
   /// Create a new [StudyProtocol].
   CAMSStudyProtocol({
-    this.studyId,
+    String ownerId,
     String name,
     String description,
-    this.owner,
+    this.studyId,
+    this.responsible,
     this.protocolDescription,
-  }) : super(ownerId: owner?.id, name: name, description: description) {
+  })
+      : super(ownerId: ownerId, name: name, description: description) {
     // TODO - move this elsewhere.... can't assumed that the programmer
     // create a protocol - s/he might download it e.g. from CARP.
     _registerFromJsonFunctions();