+++ carp_mobile_sensing/lib/domain/triggers.dart
@@ -104,7 +104,8 @@ class PeriodicTrigger extends CAMSTrigger {
     String triggerId,
     @required this.period,
     this.duration = const Duration(seconds: 1),
-  }) : super(triggerId: triggerId);
+  })
+      : super(triggerId: triggerId);
 
   Function get fromJsonFunction => _$PeriodicTriggerFromJson;
   factory PeriodicTrigger.fromJson(Map<String, dynamic> json) =>
@@ -127,7 +128,8 @@ class DateTimeTrigger extends CAMSTrigger {
     String triggerId,
     @required this.schedule,
     this.duration,
-  }) : super(triggerId: triggerId);
+  })
+      : super(triggerId: triggerId);
 
   Function get fromJsonFunction => _$DateTimeTriggerFromJson;
   factory DateTimeTrigger.fromJson(Map<String, dynamic> json) =>
@@ -326,7 +328,8 @@ class RecurrentScheduledTrigger extends PeriodicTrigger {
     //this.monthOfYear,
     this.remember = false,
     Duration duration = const Duration(seconds: 10),
-  }) : super(
+  })
+      : super(
             triggerId: triggerId,
             period: const Duration(seconds: 1),
             duration: duration) {
@@ -506,7 +509,8 @@ class CronScheduledTrigger extends CAMSTrigger {
     String triggerId,
     this.cronExpression,
     this.duration = const Duration(seconds: 1),
-  }) : super(triggerId: triggerId);
+  })
+      : super(triggerId: triggerId);
 
   static String _cronToString(
           int minute, int hour, int day, int month, int weekday) =>
@@ -540,7 +544,8 @@ class SamplingEventTrigger extends CAMSTrigger {
     @required this.measureType,
     this.resumeCondition,
     this.pauseCondition,
-  }) : super(triggerId: triggerId);
+  })
+      : super(triggerId: triggerId);
 
   /// The data type of the event to look for.
   ///
@@ -585,7 +590,7 @@ class ConditionalEvent extends Serializable {
   Map<String, dynamic> condition;
   ConditionalEvent(this.condition) : super();
 
-  operator [](String index) => condition[index];
+  dynamic operator [](String index) => condition[index];
 
   Function get fromJsonFunction => _$ConditionalEventFromJson;
   factory ConditionalEvent.fromJson(Map<String, dynamic> json) =>
@@ -619,7 +624,8 @@ class ConditionalSamplingEventTrigger extends CAMSTrigger {
     @required this.measureType,
     this.resumeCondition,
     this.pauseCondition,
-  }) : super(triggerId: triggerId);
+  })
+      : super(triggerId: triggerId);
 
   /// The data type of the event to look for.
   String measureType;
@@ -680,7 +686,8 @@ class RandomRecurrentTrigger extends CAMSTrigger {
     this.startTime,
     this.endTime,
     this.duration = const Duration(seconds: 2),
-  }) : super(triggerId: triggerId);
+  })
+      : super(triggerId: triggerId);
 
   Function get fromJsonFunction => _$RandomRecurrentTriggerFromJson;
   factory RandomRecurrentTrigger.fromJson(Map<String, dynamic> json) =>